Subject: Starting the Add Movie View Controller
Body:
Changed the name of AddWatchlistElementExtension to AddElementExtension and put in addElement functions for the Add Movie View Contoller.
Imported UIKit for the AddElementExtension file.
Added a description text view in the Add Movie View and created an outlet for that.
Set the style of the description text view.
Changed the background colour of the text views to be a custom colour I set up.